# Improve LookupContentTypeForImageUri performance in XpsFixedPageReaderWriter.cs